misc: refactor file system cache to improve performance
Some functionalities were not used and would decrease overall
performance, those have been removed.
