+++ src/Cache/FileSystemCache.php
@@ -4,17 +4,27 @@ declare(strict_types=1);
 
 namespace CuyZ\Valinor\Cache;
 
-use CuyZ\Valinor\Cache\Compiled\CompiledPhpFileCache;
-use CuyZ\Valinor\Cache\Compiled\MixedValueCacheCompiler;
+use CuyZ\Valinor\Cache\Exception\CacheDirectoryNotWritable;
+use CuyZ\Valinor\Cache\Exception\CompiledPhpCacheFileNotWritten;
+use CuyZ\Valinor\Cache\Exception\CorruptedCompiledPhpCacheFile;
 use CuyZ\Valinor\Definition\ClassDefinition;
 use CuyZ\Valinor\Definition\FunctionDefinition;
 use CuyZ\Valinor\Definition\Repository\Cache\Compiler\ClassDefinitionCompiler;
 use CuyZ\Valinor\Definition\Repository\Cache\Compiler\FunctionDefinitionCompiler;
-use Psr\SimpleCache\CacheInterface;
+use Error;
+use FilesystemIterator;
 use Traversable;
 
-use function is_object;
-use function sys_get_temp_dir;
+use function bin2hex;
+use function file_exists;
+use function file_put_contents;
+use function is_dir;
+use function mkdir;
+use function random_bytes;
+use function rename;
+use function str_contains;
+use function unlink;
+use function var_export;
 
 /**
  * @api
@@ -24,83 +34,113 @@ use function sys_get_temp_dir;
  */
 final class FileSystemCache implements WarmupCache
 {
-    /** @var array<string, CacheInterface<EntryType>> */
-    private array $delegates;
+    private const TEMPORARY_DIR_PERMISSION = 510;
 
-    public function __construct(string $cacheDir = null)
+    private const GENERATED_MESSAGE = 'Generated by ' . self::class;
+
+    private string $cacheDir;
+
+    private ClassDefinitionCompiler $classDefinitionCompiler;
+
+    private FunctionDefinitionCompiler $functionDefinitionCompiler;
+
+    public function __construct(string $cacheDir)
     {
-        $cacheDir ??= sys_get_temp_dir();
-
-        // @infection-ignore-all
-        $this->delegates = [
-            '*' => new CompiledPhpFileCache($cacheDir . DIRECTORY_SEPARATOR . 'mixed', new MixedValueCacheCompiler()),
-            ClassDefinition::class => new CompiledPhpFileCache($cacheDir . DIRECTORY_SEPARATOR . 'classes', new ClassDefinitionCompiler()),
-            FunctionDefinition::class => new CompiledPhpFileCache($cacheDir . DIRECTORY_SEPARATOR . 'functions', new FunctionDefinitionCompiler()),
-        ];
+        $this->cacheDir = $cacheDir;
+        $this->classDefinitionCompiler = new ClassDefinitionCompiler();
+        $this->functionDefinitionCompiler = new FunctionDefinitionCompiler();
     }
 
     public function warmup(): void
     {
-        foreach ($this->delegates as $delegate) {
-            if ($delegate instanceof WarmupCache) {
-                $delegate->warmup();
-            }
-        }
+        $this->createTemporaryDir();
     }
 
     public function has($key): bool
     {
-        foreach ($this->delegates as $delegate) {
-            if ($delegate->has($key)) {
-                return true;
-            }
-        }
+        $filename = $this->path($key);
 
-        return false;
+        return file_exists($filename);
     }
 
     public function get($key, $default = null): mixed
     {
-        foreach ($this->delegates as $delegate) {
-            if ($delegate->has($key)) {
-                return $delegate->get($key, $default);
-            }
+        $filename = $this->path($key);
+
+        if (! file_exists($filename)) {
+            return $default;
         }
 
-        return $default;
+        try {
+            return include $filename;
+        } catch (Error) {
+            throw new CorruptedCompiledPhpCacheFile($filename);
+        }
     }
 
     public function set($key, $value, $ttl = null): bool
     {
-        $delegate = $this->delegates['*'];
+        $filename = $this->path($key);
 
-        if (is_object($value) && isset($this->delegates[$value::class])) {
-            $delegate = $this->delegates[$value::class];
+        $code = $this->compile($value);
+
+        $tmpDir = $this->createTemporaryDir();
+
+        /** @infection-ignore-all */
+        $tmpFilename = $tmpDir . DIRECTORY_SEPARATOR . bin2hex(random_bytes(16));
+
+        try {
+            if (! @file_put_contents($tmpFilename, $code)) {
+                throw new CompiledPhpCacheFileNotWritten($tmpFilename);
+            }
+
+            if (! file_exists($filename) && ! @rename($tmpFilename, $filename)) {
+                throw new CompiledPhpCacheFileNotWritten($filename);
+            }
+        } finally {
+            if (file_exists($tmpFilename)) {
+                unlink($tmpFilename);
+            }
         }
 
-        return $delegate->set($key, $value, $ttl);
+        return true;
     }
 
     public function delete($key): bool
     {
-        $deleted = true;
+        $filename = $this->path($key);
 
-        foreach ($this->delegates as $delegate) {
-            $deleted = $delegate->delete($key) && $deleted;
+        if (file_exists($filename)) {
+            return @unlink($filename);
         }
 
-        return $deleted;
+        return true;
     }
 
     public function clear(): bool
     {
-        $cleared = true;
-
-        foreach ($this->delegates as $delegate) {
-            $cleared = $delegate->clear() && $cleared;
+        if (! is_dir($this->cacheDir)) {
+            return true;
         }
 
-        return $cleared;
+        $success = true;
+
+        /** @var FilesystemIterator $file */
+        foreach (new FilesystemIterator($this->cacheDir) as $file) {
+            if (! $file->isFile()) {
+                continue;
+            }
+
+            $line = $file->openFile()->getCurrentLine();
+
+            if (! $line || ! str_contains($line, self::GENERATED_MESSAGE)) {
+                continue;
+            }
+
+            $success = @unlink($this->cacheDir . DIRECTORY_SEPARATOR . $file->getFilename()) && $success;
+        }
+
+        return $success;
     }
 
     /**
@@ -115,13 +155,11 @@ final class FileSystemCache implements WarmupCache
 
     public function setMultiple($values, $ttl = null): bool
     {
-        $set = true;
-
         foreach ($values as $key => $value) {
-            $set = $this->set($key, $value, $ttl) && $set;
+            $this->set($key, $value, $ttl);
         }
 
-        return $set;
+        return true;
     }
 
     public function deleteMultiple($keys): bool
@@ -133,5 +171,38 @@ final class FileSystemCache implements WarmupCache
         }
 
         return $deleted;
+    }
+
+    private function compile(mixed $value): string
+    {
+        $generatedMessage = self::GENERATED_MESSAGE;
+
+        $code = match (true) {
+            $value instanceof ClassDefinition => $this->classDefinitionCompiler->compile($value),
+            $value instanceof FunctionDefinition => $this->functionDefinitionCompiler->compile($value),
+            default => var_export($value, true),
+        };
+
+        return <<<PHP
+        <?php // $generatedMessage
+        return $code;
+        PHP;
+    }
+
+    private function createTemporaryDir(): string
+    {
+        $tmpDir = $this->cacheDir . DIRECTORY_SEPARATOR . '.valinor.tmp';
+
+        if (! is_dir($tmpDir) && ! @mkdir($tmpDir, self::TEMPORARY_DIR_PERMISSION, true)) {
+            throw new CacheDirectoryNotWritable($this->cacheDir);
+        }
+
+        return $tmpDir;
+    }
+
+    private function path(string $key): string
+    {
+        /** @infection-ignore-all */
+        return $this->cacheDir . DIRECTORY_SEPARATOR . $key . '.php';
     }
 }
